# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Storecipe public API contract
  version: 0.3.0
  description: >-
    Public contract for recipes, ratings, deterministic recipe queries, and imports. The internal
    `POST /internal/recipes/source-lookup` endpoint is M2M-only and is excluded
    from the public proxy.
servers:
  - url: https://api.storecipe.example
security:
  - bearerAuth: []
paths:
...
  /v1/imports:
    get:
      operationId: listImports
      description: Requires `recipes:read`; lists only retained owner metadata.
      parameters:
        - {name: cursor, in: query, schema: {type: [string, "null"], maxLength: 1024}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        "200":
          description: Retained import history
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportHistoryPage"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "410": {$ref: "#/components/responses/AccountDeleted"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /v1/imports/url:
    post:
      operationId: importRecipeFromUrl
      description: >-
        Requires `recipes:write`. A matching idempotency replay returns the current
        job. The default Catalog source lookup rejects a saved recipe source with
        `recipe_source_exists` when `duplicatePolicy` is `warn`; `allow` permits
        that saved-recipe duplicate but cannot bypass an active URL import job.
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UrlImportRequest"}
      responses:
        "202":
          description: New durable import accepted
          headers:
            Location: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportAccepted"}
        "200":
          description: Existing idempotent import
          headers:
            Location: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportJob"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "410": {$ref: "#/components/responses/AccountDeleted"}
        "409":
          description: >-
            `active_url_import_exists` includes `existingJobId`; `recipe_source_exists`
            includes `existingRecipeId`.
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
        "429": {$ref: "#/components/responses/ImportBurstExceeded"}
        "413": {$ref: "#/components/responses/RequestTooLarge"}
        "503":
          description: >-
            Catalog source lookup is unavailable, or Redis cannot make an import
            admission decision (`rate_limit_unavailable`).
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/Problem"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /v1/imports/text:
    post:
      operationId: importRecipeFromText
      description: Requires `recipes:write`. Text is limited to 256 KiB UTF-8.
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TextImportRequest"}
      responses:
        "202":
          description: New durable import accepted
          headers:
            Location: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportAccepted"}
        "200":
          description: Existing idempotent import
          headers:
            Location: {schema: {type: string}}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportJob"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "410": {$ref: "#/components/responses/AccountDeleted"}
        "409": {$ref: "#/components/responses/Conflict"}
        "429": {$ref: "#/components/responses/ImportBurstExceeded"}
        "413": {$ref: "#/components/responses/RequestTooLarge"}
        "503": {$ref: "#/components/responses/RateLimitUnavailable"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /v1/imports/{jobId}:
    parameters:
      - {name: jobId, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getImportStatus
      description: Requires `recipes:read`; another user's job is reported as not found.
      responses:
        "200": {description: Import job status, content: {application/json: {schema: {$ref: "#/components/schemas/ImportJob"}}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
    delete:
      operationId: cancelImport
      description: Requires `recipes:write`; cancellation is cooperative once processing starts.
      responses:
        "204": {description: Queued or already-cancelled job is cancelled}
        "202":
          description: Cooperative cancellation requested
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportJob"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /v1/imports/{jobId}/draft:
    parameters:
      - {name: jobId, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getImportDraft
      description: >-
        Requires `recipes:read`. Returns the extracted recipe retained for a
        `review_required` job owned by the caller. Missing jobs, other owners,
        and jobs without a retained extract are reported as not found.
      responses:
        "200":
          description: Extracted recipe draft
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportReviewDraft"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 1, maxLength: 255}
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      headers:
        WWW-Authenticate:
          schema: {type: string}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Forbidden:
      description: Access token lacks a required scope
      headers:
        WWW-Authenticate:
          schema: {type: string}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    AccountDeleted:
      description: The authenticated subject is protected by a deletion tombstone.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
          example:
            type: https://docs.storecipe.example/problems/account_deleted
            title: Gone
            status: 410
            detail: This account has been deleted.
            request_id: 01JEXAMPLEACCOUNTDELETED
            errorCategory: account_deleted
    NotFound:
      description: Resource not found or not owned by the authenticated user
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    Conflict:
      description: Request conflicts with durable resource state
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    ImportBurstExceeded:
      description: Import submission burst limit exceeded; retry after the indicated interval.
      headers:
        Retry-After:
          schema: {type: integer, minimum: 1}
        RateLimit-Limit:
          schema: {type: integer, minimum: 1}
        RateLimit-Remaining:
          schema: {type: integer, minimum: 0}
        RateLimit-Reset:
          schema: {type: integer, minimum: 0}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    RateLimitUnavailable:
      description: Redis could not make a rate-limit admission decision; retry after the indicated interval.
      headers:
        Retry-After:
          schema: {type: integer, minimum: 1}
        RateLimit-Limit:
          schema: {type: integer, minimum: 1}
        RateLimit-Remaining:
          schema: {type: integer, minimum: 0}
        RateLimit-Reset:
          schema: {type: integer, minimum: 0}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    RequestTooLarge:
      description: Request body exceeds the service size limit.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
  schemas:
...
    ImportAccepted:
      type: object
      required: [jobId, status]
      properties:
        jobId: {type: string, format: uuid}
        status: {const: queued}
    UrlImportRequest:
      type: object
      required: [url]
      properties:
        url: {type: string, format: uri, maxLength: 2048}
        duplicatePolicy:
          type: string
          enum: [warn, allow]
          default: warn
    TextImportRequest:
      type: object
      required: [text]
      properties:
        text: {type: string, minLength: 1, maxLength: 262144}
    ImportJob:
      type: object
      required: [id, status, attemptCount, createdRecipeId, errorCategory, cancellationRequested, hasCandidate]
      properties:
        id: {type: string, format: uuid}
        status: {enum: [queued, processing, completed, review_required, failed, cancelled, timed_out]}
        attemptCount: {type: integer, minimum: 0}
        createdRecipeId: {type: [string, "null"], format: uuid}
        errorCategory: {type: [string, "null"]}
        cancellationRequested: {type: boolean}
        hasCandidate: {type: boolean}
    ImportHistoryItem:
      type: object
      required: [id, inputKind, createdAt, updatedAt, terminalAt, status, phase]
      properties:
        id: {type: string, format: uuid}
        inputKind: {type: string, enum: [url, text]}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        terminalAt: {type: [string, "null"], format: date-time}
        status: {type: string, enum: [queued, processing, completed, review_required, failed, cancelled, timed_out]}
        phase: {type: string, enum: [waiting, fetching, extracting, validating, saving, completed, review_required, failed, cancelled, timed_out]}
    ImportHistoryPage:
      type: object
      required: [items, nextCursor]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/ImportHistoryItem"}}
        nextCursor: {type: [string, "null"], maxLength: 1024}
    ImportReviewDraft:
      type: object
      required: [ingredients, instructions]
      properties:
        title: {type: [string, "null"]}
        sourceUrl: {type: [string, "null"]}
        servings: {type: [integer, "null"]}
        prepMinutes: {type: [integer, "null"]}
        cookMinutes: {type: [integer, "null"]}
        totalMinutes: {type: [integer, "null"]}
        ingredients:
          type: array
          items: {type: string}
        instructions:
          type: array
          items: {type: string}
        tags:
          type: array
          items: {type: string}
...
    Problem:
      type: object
      required: [type, title, status, request_id]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        request_id: {type: string}
        errorCategory: {type: string}
        existingJobId: {type: string, format: uuid}
        existingRecipeId: {type: string, format: uuid}